+++ test/test_transaction_utils.h
@@ -1,0 +1,7 @@
+#include "../lib/transaction.h"
+#include "../lib/txin.h"
+#include "../lib/txout.h"
+
+TxIn CreateFakeTxIn();
+TxOut CreateFakeTxOut();
+Transaction CreateFakeTransaction();
